08/05

+++ lib/models/jsontempo.dart
@@ -177,7 +177,7 @@ class Main {
 }
 
 class Wind {
-  int? speed;
+  double? speed;
   int? deg;
   double? gust;
